+++ frontend/src/services/driver.expense.service.js
@@ -4,14 +4,14 @@ const driverExpenseService = {
     createExpense(expesneData){
         return api.post('/expense/driver/create',expesneData)
     },
-    getExpenses(){
-        return api.get('/expesne/driver/get-expenses')
+    getExpenses(userId){
+        return api.get('/expesne/driver/get-expenses',{ params: { userId }})
     },
     updateExpense(id,updatedData){
         return api.put(`/expense/driver/update-expense`,updatedData)
     },
-    deleteExpense(id){
-        return api.delete(`/expense/driver/delete-expense/${id}`)
+    deleteExpense(id,userId){
+        return api.delete(`/expense/driver/delete-expense/${id}`,{ params: { userId }})
     }
 }
