+++ frontend/src/services/vehicle.expense.service.js
@@ -1,17 +1,17 @@
 import api from "../libraries/api"
 
 const vehicleExpenseService = {
-    createExpense(expesneData){
-        return api.post('/expense/vehicle/create',expesneData)
+    createExpense(expenseData){
+        return api.post('/expense/vehicle/create',expenseData)
     },
-    getExpenses(){
+    getExpenses(userId){
         return api.get('/expesne/vehicle/get-expenses')
     },
     updateExpense(id,updatedData){
         return api.put(`/expense/vehicle/update-expense`,updatedData)
     },
-    deleteExpense(id){
-        return api.delete(`/expense/vehicle/delete-expense/${id}`)
+    deleteExpense(id,userId){
+        return api.delete(`/expense/vehicle/delete-expense/${id}`,{ params: { userId }})
     }
 }
